Refactor solution to single CommEx extension plugin
